pyproject.toml
[project]
name = "aisafetyeventsandtraining"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Orpheus Lummis", email = "o@orpheuslummis.info" }
]
dependencies = [
    "exa-py>=1.0.12",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/aisafetyeventsandtraining"]
